Move booking arrivals Api endpoint to CAS3 namespace
